extracted custom fs
